Combustion cap
If you somehow get your combustion from different sources to stack above 80, it will act as if it was max of 80.

Because if it goes above 100, you can't ever trigger combustion.
And even a 100 is hard to get in some ocassions since it's the max and update tick weirdness

approved by perdition